requirements.txt
aiohttp==3.7.4.post0
appnope==0.1.2
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
debugpy==1.5.1
decorator==5.1.1
discord.py==1.7.3
distlib==0.3.4
dnspython==2.2.0
entrypoints==0.4
executing==0.8.2
filelock==3.5.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
matplotlib-inline==0.1.3
mongomock==4.0.0
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2022.1.8
platformdirs==2.5.0
pluggy==1.0.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
Pygments==2.11.2
pymongo==4.0.1
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pyzmq==22.3.0
requests==2.27.1
sentinels==1.0.0
six==1.16.0
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.13.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
yarl==1.7.2
